refactor: provider & context 분리
