+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { NotificationProvider } from "./provider/NotificationProvider";
 import { UserProvider } from "./provider/UserProvider";
-import { ToggleProvider } from "./provider/ToggleProvider";
+import { ThemeProvider } from "./provider/ThemeProvider";
 import { Home } from "./components/Home";
 
 // 메인 App 컴포넌트
@@ -8,9 +8,9 @@ const App: React.FC = () => {
   return (
     <NotificationProvider>
       <UserProvider>
-        <ToggleProvider>
+        <ThemeProvider>
           <Home />
-        </ToggleProvider>
+        </ThemeProvider>
       </UserProvider>
     </NotificationProvider>
   );
