+++ src/context/useNotificationContext.ts
@@ -1,0 +1,21 @@
+import { createContext, useContext } from "react";
+import { Notification } from "../type";
+
+interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
+
+export const useNotificationContext = () => {
+  const context = useContext(NotificationContext);
+
+  if (!context) {
+    throw new Error("NotificationContext 사용이 불가능해요...");
+  }
+  return context;
+};
