+++ src/provider/NotificationProvider.tsx
@@ -1,26 +1,7 @@
-import { createContext, useContext, useState } from "react";
-import { Notification } from "../type";
+import { useState } from "react";
 import { useCallback, useMemo } from "../@lib";
-
-interface NotificationContextType {
-  notifications: Notification[];
-  addNotification: (message: string, type: Notification["type"]) => void;
-  removeNotification: (id: number) => void;
-}
-
-const NotificationContext = createContext<NotificationContextType | undefined>(
-  undefined
-);
-
-// eslint-disable-next-line react-refresh/only-export-components
-export const useNotificationContext = () => {
-  const context = useContext(NotificationContext);
-  if (!context) {
-    throw new Error("노티파이 사용 불가!");
-  }
-
-  return context;
-};
+import { Notification } from "../type";
+import { NotificationContext } from "../context/useNotificationContext";
 
 export const NotificationProvider = ({
   children,
