+++ src/provider/ThemeProvider.tsx
@@ -1,0 +1,25 @@
+import { useState } from "react";
+import { useCallback, useMemo } from "../@lib";
+import { ThemeContext } from "../context/useThemeContext";
+
+export const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  }, []);
+
+  const contextValue = useMemo(
+    () => ({
+      toggleTheme,
+      theme,
+    }),
+    [theme, toggleTheme]
+  );
+
+  return (
+    <ThemeContext.Provider value={contextValue}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
